Began adding and testing Dispatch Unit
Added ability to flush the cache. This is currently useful for testing output values, but likely will be crucial when propagating texture writes back to the host.
Fixed some bugs in the cache.
Began adding a dispatch unit. This will decode instructions in one cycle, then execute the instruction potentially multiple times over several cycles. This primarily manifests in 2 ways: 1.) vector ops are done using SIMT, thus we can only dispatch to one FP core per sub clock cycle; and 2.) we replicate operations across 4 pixels at a time when running the fragment fragment shader.
Currently there are 2 dispatch units per SM, and they each get 3 sub clock cycles per clock cycle. This allows up to 6 instructions to be dispatched per clock cycle. This value may be lower if there is contestion or if a sub clock cycle is dedicated to a decode.
If a dispatch unit gets stalled in a sub clock cycle waiting for a resource to be ready it will stop executing for the remainder of the clock cycle.