Added Pipelining
Changed microcode to use full width register addressing (12 bits) as opposed to a 256 register view.
This has a variety of benefits, one of the key ones is that the SM no longer queries the Dispatch Unit to get the base register, this allows the Dispatch Unit to switch thread context (warp) and have a pipelined processor access the correct register instead of accessing a register in the new view.
Added pipelining to the FPU. There are 5 cycles, Register Read, Read Lock Release, Operation, Register Write, Write Lock Release.
Added debugger. This currently allows stepping clock cycles, and viewing registers. Register viewing is done by selecting the SM, DU, and Replication Index. This is still incredibly buggy due to locking reads/writes with the pipes. For now it works well enough to have debugged why pipelining wasn't working.
Added a debug check to log when a register unlock has underflowed.