Updated VBox headers, added ROM, mapped memory
Updated the VBox headers to the latest version.
Added an Expansion ROM. This only sort of works right now, the UEFI driver is loaded, but not run.
Added Link registers to the PCIe capability, this was required and was being incorrectly parsed previously.
Fixed memory mapping; the PDM was performing special handling to update internal mappings, I've copied that code into my device now.
Added DllMain, this wasn't actually necessary, there was a problem loading the device, but that was from the new VBox API changing the RC error stringification function changing.