Made more like hardware
Changed a lot of the currently functional paths to be more like hardware. This means they execute over multiple clock cycles and use bus objects. This is incredibly slow, but makes it much easier to translate to hardware.
Added display manager for producing EDID data for the virtual display, as well as configuring its size.