Minor patches to logging and rendering
Added mutex locking to logging
Fixed an NPE caused during disabled displays, this was because I was `memcpy`'ing from `0`, rather than `memset`'ing to `0`.
Locked framerate to ~60 FPS. This was mostly reduce the CPU and GPU load. Eventually I'd like to have the framerate programmable by the system, so that it can support for VBE modes.
Disabled user window resizing and fullscreening. The window size is set by the OS, and the framebuffer is just a raw linear buffer, so a user changing the size causes slipping of the scanlines.