Improved VHDL macros
There are a lot of changes in this one.
First, some helpers for process calls were added. These had to be added because some processes could trigger changes that necessitated other processes. This could create an infinite loop that would stall the program.
Second, the Sensitivity enum is now defined with a macro, thus allows us to ensure it is always named correctly.
Third, the names of the enum members should now be the actual names of the variables. This will make more sense shortly, but it makes it easier to see the analogue VHDL code.
With the change to the validating triggers for processes it also became advantageous to have some more macros.
The `PROCESSES_DECL` macro now declares the generic `Processes` function that most entities would have. This allows us to ensure that the function and the `trigger` parameter are correctly named.
Processes are now declared with their own macro: `PROCESS_DECL`, this again ensures that the `trigger` parameter is correctly named.
Calling the `Processes` is now done with a macro: `TRIGGER_SENSITIVTY`. This has 2 purposes: first, the name of `Processes` is now opaque to the entity; second we can now omit the `Sensitivity` enum name. This is where naming the enum members the same as the entity members starts to become advantageous. It will now be much clearer what is being triggered.
A similar effect is now had for the `RisingEdge` and `FallingEdge` functions. Where before you had to pass the enum name, the variable, and the trigger sense, we can now simply call `RISING_EDGE(var)`. Because the name for `trigger` is now opaque, we can guarantee it is always what we want, and because the enum members and entity members are named identically, it is trivial to pass them both with a single macro parameter.
There are also some equivalent functions for when the signal is actually mux. The only difference is this calls the signal mux function rather than passing it as a variable.
Entering Processes from the `Processes` function is now simplified. With these changes it became necessary to always wrap the call in an if to check for the triggers, now we can do that with a single macros that will call the Process if the trigger was correct. This unfortunately requires a lot of helper macros, but at least it works.
As much as I don't like macros, these make it much quicker to write out code (eliminating a lot of boilerplate chars), as well as easing understanding since you can now clearly see the relation between a signal and how it triggers a process.