Reworked DCFIFO
Rather than using getters, I switched to using callback functions. This changes out variables to be pushed instead of pulled. This is particularly important for hardware, since an out variable can be used as part of  a mux or process.

Also reworked the design a bit to be more inline with the paper's Verilog, rather than some other VHDL I was using as a partial reference.

It also doesn't work, but that's a separate matter.