Cleaned up DCFIFO TestBench
With the fixes applied to the ReadPointer and WritePointer, I cleaned up the TestBench for the DCFIFO, and now test at a variety of bit counts. The paper only mentioned up to 3 address bits, and had some unclear descriptions on what would happen over 3 bits, so I test up to 5 bits. Fortunately it continues working fine.