Worked on runner
Moved out some of the legacy tests.
Fixed a Vulkan bug regarding semaphore usage.
Finally fixed the screen being black: I added functionality a while ago to blackout the screen when the GPU was supposed to be inactive, to make it easier to not update code I set that parameter to a default value. Setting it to be active in SoftGpuRunner fixed the issue.